#355
- Removed Duplicate Menu Action Plugin
